+++ Minibank.Core/ICurrencyConverter.cs
@@ -8,6 +8,6 @@
         /// <param name="amount">Currency amount in rubles</param>
         /// <param name="fromCurrency">Currency code to be transferred</param>
         /// <param name="toCurrency">Currency code to which you want to transfer</param>
-        double ConvertCurrency(int amount, string fromCurrency, string toCurrency);
+        double ConvertCurrency(double amount, string fromCurrency, string toCurrency);
     }
 }
